web/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.5.0; python_full_version >= '3.6.2'
asgiref==3.5.0; python_version >= '3.7'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3.5'
click==8.1.2; python_version >= '3.7'
django-better-admin-arrayfield==1.4.2
django-filter==21.1
django-formtools==2.3; python_version >= '3.6'
django-google-maps==0.13.0
django-otp==1.1.3
django-phonenumber-field==6.1.0; python_version >= '3.7'
django-two-factor-auth[phonenumbers]==1.13.2
django==4.0.4
djangorestframework==3.13.1
drf-spectacular==0.22.1
gunicorn==20.1.0
h11==0.12.0; python_version >= '3.6'
httpcore==0.14.7; python_version >= '3.6'
httpx==0.22.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
importlib-metadata==4.11.3; python_version < '3.10'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.4.0; python_version >= '3.7'
markdown==3.3.6
messagebird==2.0.0
numpy==1.22.3
onesignal-sdk==2.0.0
phonenumbers==8.12.47
psycopg2-binary==2.9.3
pyrsistent==0.18.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-liquid==1.1.7
pytz==2022.1
pyyaml==6.0; python_version >= '3.6'
qrcode==7.3.1; python_version >= '3.6'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rfc3986[idna2008]==1.5.0
sentry-sdk==1.5.10
setuptools==62.1.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlparse==0.4.2; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.17.6
whitenoise==6.0.0
zipp==3.8.0; python_version >= '3.7'
